requirements.txt
pushbullet.py~=0.12.0

requests~=2.26.0